bugfix - getValues doesn't call super.
This results in getValues not being called further up the stack (hof-form-controller/lib/controller)
